+++ components/common/buttonBig.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 import React from 'react';
-import media from '../libs/client/media';
+import media from '../../libs/client/media';
 
 interface ButtonProps {
   children: string | React.ReactNode;
@@ -8,21 +8,26 @@ interface ButtonProps {
   [key: string]: any;
 }
 
-const ButtonBig = ({ children, active = true }: ButtonProps) => {
-  return <ButtonWrapper active={active}>{children}</ButtonWrapper>;
+const ButtonBig = ({ children, disabled = false }: ButtonProps) => {
+  return <ButtonWrapper disabled={disabled}>{children}</ButtonWrapper>;
 };
 
 export default ButtonBig;
 
-const ButtonWrapper = styled.button<{ active: boolean }>`
+const ButtonWrapper = styled.button`
   margin-bottom: 12px;
   padding: 12px;
   width: 100%;
   max-width: 350px;
   ${media.tablet`max-width:500px`};
-  border: 1px solid black;
-  background-color: ${(props) => (props.active ? 'black' : 'gray')};
   color: white;
+  border: 1px solid black;
+  background-color: black;
   font-size: 14px;
   cursor: pointer;
+
+  &:disabled {
+    background-color: gray;
+    border: 1px solid gray;
+  }
 `;
